+++ musicbot/commands/botmanipulate.py
@@ -1,31 +1,15 @@
 import logging
 import discord
 import aiohttp
-from functools import wraps
 
 from ..utils import _get_variable
 from .. import exceptions
 from ..constructs import Response
+from ..wrappers import owner_only
 
 log = logging.getLogger(__name__)
 
 cog_name = 'botmanipulate'
-
-# @TheerapakG: TODO: move wrappers into one place
-# TODO: Add some sort of `denied` argument for a message to send when someone else tries to use it
-def owner_only(func):
-    @wraps(func)
-    async def wrapper(bot, *args, **kwargs):
-        # Only allow the owner to use these commands
-        orig_msg = _get_variable('message')
-
-        if not orig_msg or orig_msg.author.id == bot.config.owner_id:
-            # noinspection PyCallingNonCallable
-            return await func(bot, *args, **kwargs)
-        else:
-            raise exceptions.PermissionsError("Only the owner can use this command.", expire_in=30)
-
-    return wrapper
 
 async def cmd_disconnect(bot, guild):
     """
